ADD injecting payload into executable:
First, we modify the deadbeef and cafebabe tmp variables into __bytecode with the corresponding offsets and size.
Then, we inject the final __bytecode into the executable file, with a new entrypoint and in the same time, we add a signature into elf header from executable file to know we already packed it.